+++ tokenizers/src/tokenizer/added_vocabulary.rs
@@ -92,7 +92,7 @@ impl std::hash::Hash for AddedToken {
     }
 }
 
-type MatchingSet = (AhoCorasick, Vec<u32>);
+type MatchingSet = (AhoCorasick, Vec<u64>);
 
 lazy_static! {
     static ref STARTS_WITH_WORD: Regex = Regex::new(r"^\w").unwrap();
@@ -142,10 +142,10 @@ fn space_rightmost_at_start(sentence: &str) -> usize {
 pub(super) struct AddedVocabulary {
     /// Contains the mapping from String (token content) to ID. This map contains both special
     /// tokens and classic added tokens that were added to the this vocabulary.
-    added_tokens_map: HashMap<String, u32>,
+    added_tokens_map: HashMap<String, u64>,
     /// Contains the mapping from ID to AddedToken for all the added tokens, both special
     /// and classic.
-    added_tokens_map_r: HashMap<u32, AddedToken>,
+    added_tokens_map_r: HashMap<u64, AddedToken>,
 
     /// Contains only the classic AddedToken, in the specific order the user gave them.
     added_tokens: Vec<AddedToken>,
@@ -187,17 +187,17 @@ impl AddedVocabulary {
     }
 
     /// Get the additional vocabulary
-    pub fn get_vocab(&self) -> &HashMap<String, u32> {
+    pub fn get_vocab(&self) -> &HashMap<String, u64> {
         &self.added_tokens_map
     }
 
     /// Get the additional vocabulary with the AddedTokens
-    pub fn get_added_tokens_decoder(&self) -> &HashMap<u32, AddedToken> {
+    pub fn get_added_tokens_decoder(&self) -> &HashMap<u64, AddedToken> {
         &self.added_tokens_map_r
     }
 
     /// Get the id matching one of our token if it exists
-    pub fn token_to_id(&self, token: &str, model: &impl Model) -> Option<u32> {
+    pub fn token_to_id(&self, token: &str, model: &impl Model) -> Option<u64> {
         self.added_tokens_map
             .get(token)
             .copied()
@@ -205,7 +205,7 @@ impl AddedVocabulary {
     }
 
     /// Get the token matching the given id if it exists
-    pub fn id_to_token(&self, id: u32, model: &impl Model) -> Option<String> {
+    pub fn id_to_token(&self, id: u64, model: &impl Model) -> Option<String> {
         self.added_tokens_map_r
             .get(&id)
             .map(|t| t.content.clone())
@@ -258,12 +258,12 @@ impl AddedVocabulary {
                 new_id
             } else {
                 self.added_tokens_map.values().cloned().max().map_or(
-                    model.get_vocab_size() as u32,
+                    model.get_vocab_size() as u64,
                     |max| {
-                        if (max >= model.get_vocab_size() as u32) || model.get_vocab_size() == 0 {
+                        if (max >= model.get_vocab_size() as u64) || model.get_vocab_size() == 0 {
                             max + 1
                         } else {
-                            model.get_vocab_size() as u32
+                            model.get_vocab_size() as u64
                         }
                     },
                 )
@@ -297,7 +297,7 @@ impl AddedVocabulary {
     /// We keep two different RegexSet, one that will take care of matching against the
     /// non-normalized string, and one matching against the normalized one.
     fn refresh_added_tokens<N: Normalizer>(&mut self, model: &impl Model, normalizer: Option<&N>) {
-        type TupleTokenId<'a> = (&'a AddedToken, u32);
+        type TupleTokenId<'a> = (&'a AddedToken, u64);
         let (normalized, non_normalized): (Vec<TupleTokenId>, Vec<TupleTokenId>) = self
             .special_tokens
             .iter()
@@ -311,13 +311,13 @@ impl AddedVocabulary {
             })
             .partition(|(token, _)| token.normalized);
 
-        let (tokens, ids): (Vec<&AddedToken>, Vec<u32>) = non_normalized.into_iter().unzip();
+        let (tokens, ids): (Vec<&AddedToken>, Vec<u64>) = non_normalized.into_iter().unzip();
         let trie = AhoCorasickBuilder::new()
             .match_kind(MatchKind::LeftmostLongest)
             .build(tokens.iter().map(|token| &token.content));
         self.split_trie = (trie, ids);
 
-        let (ntokens, nids): (Vec<&AddedToken>, Vec<u32>) = normalized.into_iter().unzip();
+        let (ntokens, nids): (Vec<&AddedToken>, Vec<u64>) = normalized.into_iter().unzip();
         let patterns: Vec<_> = ntokens
             .iter()
             .map(|token| {
@@ -338,7 +338,7 @@ impl AddedVocabulary {
     /// This method returns a list "splits", each of them being a pair of Offsets
     /// and an optional ID if it is an AddedToken.
     /// The list of splits cover the entire input string.
-    fn find_matches(&self, sentence: &str, split_re: &MatchingSet) -> Vec<(Option<u32>, Offsets)> {
+    fn find_matches(&self, sentence: &str, split_re: &MatchingSet) -> Vec<(Option<u64>, Offsets)> {
         if sentence.is_empty() {
             return vec![(None, (0, 0))];
         }
@@ -447,7 +447,7 @@ impl AddedVocabulary {
 #[derive(Debug, Serialize, Deserialize)]
 pub(super) struct AddedTokenWithId {
     /// The id assigned to this token
-    pub id: u32,
+    pub id: u64,
     #[serde(flatten)]
     /// The target AddedToken
     pub token: AddedToken,
@@ -488,15 +488,15 @@ mod tests {
 
     #[derive(Serialize, Deserialize)]
     struct ModelMock {
-        vocab: HashMap<String, u32>,
-        vocab_r: HashMap<u32, String>,
+        vocab: HashMap<String, u64>,
+        vocab_r: HashMap<u64, String>,
     }
     impl ModelMock {
         pub fn new<I>(iter: I) -> Self
         where
-            I: IntoIterator<Item = &'static (&'static str, u32)>,
+            I: IntoIterator<Item = &'static (&'static str, u64)>,
         {
-            let vocab: HashMap<String, u32> = iter
+            let vocab: HashMap<String, u64> = iter
                 .into_iter()
                 .map(|&(tok, id)| (tok.to_string(), id))
                 .collect();
@@ -510,7 +510,7 @@ mod tests {
         }
     }
 
-    fn simplify_output(result: &'_ PreTokenizedString) -> Vec<(&'_ str, Option<Vec<u32>>)> {
+    fn simplify_output(result: &'_ PreTokenizedString) -> Vec<(&'_ str, Option<Vec<u64>>)> {
         result
             .get_splits(OffsetReferential::Original, OffsetType::Byte)
             .into_iter()
@@ -550,13 +550,13 @@ mod tests {
         fn tokenize(&self, _sequence: &str) -> Result<Vec<Token>> {
             unimplemented!()
         }
-        fn token_to_id(&self, token: &str) -> Option<u32> {
+        fn token_to_id(&self, token: &str) -> Option<u64> {
             self.vocab.get(token).copied()
         }
-        fn id_to_token(&self, id: u32) -> Option<String> {
+        fn id_to_token(&self, id: u64) -> Option<String> {
             self.vocab_r.get(&id).cloned()
         }
-        fn get_vocab(&self) -> HashMap<String, u32> {
+        fn get_vocab(&self) -> HashMap<String, u64> {
             self.vocab.clone()
         }
         fn get_vocab_size(&self) -> usize {
